# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SKF Racing Hub API",
    "description": "REST API of the SKF Racing Hub site.\n\n**Authentication.** A `session_id` cookie, set by the Discord login. Operations\nthat need a session list it under *security*; the roles they accept are in\n`x-required-roles`.\n\n**Errors.** Every error is an `application/problem+json` document (`Problem`).\n`code` is a stable machine-readable identifier, `detail` a message that can be\nshown to the user. Validation failures (422) also carry `errors`, one entry per\nrejected field.\n\n**Paging.** Collections that take `limit` and `offset` return a plain array;\nthe total is in `X-Total-Count` and neighbouring pages in an RFC 8288 `Link`\nheader.\n\n**Tracing.** Every response carries `X-Request-ID`; quote it when reporting a\nproblem.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/regulations": {
      "get": {
        "tags": [
          "Regulations"
        ],
        "summary": "List regulations",
        "description": "Visible pages with their title in *lang* (falling back to any language).",
        "operationId": "listRegulations",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 10,
              "pattern": "^[A-Za-z0-9_-]+$",
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegulationPageSummaryOut"
                  }
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        }
      }
    },
    "/api/v1/regulations/{slug}": {
      "get": {
        "tags": [
          "Regulations"
        ],
        "summary": "Get regulation",
        "operationId": "getRegulation",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 10,
              "pattern": "^[A-Za-z0-9_-]+$",
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulationContentOut"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        }
      }
    },
    "/api/v1/regulation-pages": {
      "get": {
        "tags": [
          "Regulations"
        ],
        "summary": "List regulation pages",
        "operationId": "listRegulationPages",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RegulationPageOut"
                  },
                  "type": "array"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      },
      "post": {
        "tags": [
          "Regulations"
        ],
        "summary": "Create regulation page",
        "operationId": "createRegulationPage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulationPageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulationPageOut"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "Path of the created resource.",
                "schema": {
                  "type": "string",
                  "format": "uri-reference"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      }
    },
    "/api/v1/regulation-pages/{pageId}": {
      "get": {
        "tags": [
          "Regulations"
        ],
        "summary": "Get regulation page",
        "operationId": "getRegulationPage",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulationPageOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      },
      "patch": {
        "tags": [
          "Regulations"
        ],
        "summary": "Update regulation page",
        "operationId": "updateRegulationPage",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegulationPageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegulationPageOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      },
      "delete": {
        "tags": [
          "Regulations"
        ],
        "summary": "Delete regulation page",
        "operationId": "deleteRegulationPage",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "operationId": "listUsers",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  }
                }
              }
            },
            "headers": {
              "X-Total-Count": {
                "description": "Size of the whole collection, across all pages.",
                "schema": {
                  "type": "integer"
                }
              },
              "Link": {
                "description": "RFC 8288 links to the `next` and `prev` pages, when they exist.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      }
    },
    "/api/v1/users/{userId}": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "operationId": "updateUser",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      }
    },
    "/api/v1/users/{userId}/sessions": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Revoke user sessions",
        "description": "Delete all sessions of a user (force logout).\n\nRequires role: `admin`, `super_admin`.",
        "operationId": "revokeUserSessions",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ]
      }
    },
    "/api/v1/users/{userId}/managed-communities": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user managed communities",
        "operationId": "getUserManagedCommunities",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedCommunitiesOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ],
        "description": "Requires role: `admin`, `super_admin`."
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Set user managed communities",
        "description": "Replace the full set of communities the user manages.\n\nRequires role: `admin`, `super_admin`.",
        "operationId": "setUserManagedCommunities",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedCommunitiesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedCommunitiesOut"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableContent"
          }
        },
        "x-required-roles": [
          "admin",
          "super_admin"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "ManagedCommunitiesOut": {
        "properties": {
          "communityIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array"
          }
        },
        "type": "object",
        "required": [
          "communityIds"
        ]
      },
      "ManagedCommunitiesUpdate": {
        "properties": {
          "communityIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array"
          }
        },
        "type": "object",
        "required": [
          "communityIds"
        ]
      },
...
      "Problem": {
        "description": "An RFC 7807 problem document. Every error response has this shape.",
        "example": {
          "code": "not_found",
          "detail": "Driver not found.",
          "status": 404,
          "title": "Not Found",
          "type": "about:blank"
        },
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "type": "object"
      },
...
      "RegulationContentOut": {
        "properties": {
          "lang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "lang",
          "title",
          "subtitle",
          "content"
        ]
      },
      "RegulationContentUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300
          },
          "subtitle": {
            "type": "string",
            "maxLength": 500,
            "default": ""
          },
          "content": {
            "type": "string",
            "maxLength": 200000,
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title"
        ]
      },
      "RegulationPageCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "sortOrder": {
            "type": "integer",
            "default": 0
          },
          "isVisible": {
            "type": "boolean",
            "default": true
          },
          "contents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RegulationContentUpdate"
            },
            "type": "object",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "slug"
        ]
      },
      "RegulationPageOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer"
          },
          "isVisible": {
            "type": "boolean"
          },
          "contents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RegulationContentOut"
            },
            "type": "object"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "sortOrder",
          "isVisible",
          "contents"
        ]
      },
      "RegulationPageSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer"
          },
          "isVisible": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "sortOrder",
          "isVisible",
          "title"
        ]
      },
      "RegulationPageUpdate": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "sortOrder": {
            "type": "integer"
          },
          "isVisible": {
            "type": "boolean"
          },
          "contents": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RegulationContentUpdate"
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "discordId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "discordNickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nickname on the SKF Discord server."
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "blocked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "driverId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ]
          },
          "managedCommunityIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "discordId",
          "username",
          "displayName",
          "discordNickname",
          "avatarUrl",
          "role",
          "blocked",
          "createdAt",
          "lastLoginAt",
          "driverId",
          "managedCommunityIds"
        ]
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "driver",
          "moderator",
          "racing_judge",
          "community_manager",
          "admin",
          "super_admin"
        ]
      },
      "UserUpdate": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "blocked": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ValidationProblem": {
        "description": "A 422: the request did not pass validation.",
        "example": {
          "code": "validation_error",
          "detail": "year: Field required",
          "errors": [
            {
              "field": "year",
              "message": "Field required"
            }
          ],
          "status": 422,
          "title": "Unprocessable Content",
          "type": "about:blank"
        },
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "errors"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session id, set as a cookie by the Discord login.",
        "in": "cookie",
        "name": "session_id"
      },
      "ingestToken": {
        "type": "http",
        "description": "Shared token of the external incident parsers.",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "UnprocessableContent": {
        "description": "Unprocessable Content",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationProblem"
            }
          }
        }
      }
    }
  },
...
}